requirements.txt
# API y HTTP
requests==2.31.0
urllib3==2.1.0

# Data processing
pandas==2.1.4
openpyxl==3.1.2

# Configuration
python-dotenv==1.0.0

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# Utilities
colorama==0.4.6

# Web Framework
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1

# Security
Flask-Bcrypt==1.0.1

# Real-time WebSockets
Flask-SocketIO==5.3.5
python-socketio==5.10.0
eventlet==0.33.3
